refactor(runner): dedupe npm-install spawn into canonical helper
